feat: changement de fenêtre une fois que l'utilisateur est connecté
