+++ src/main/java/com/cytech/trashtreasure/application/PrimaryStageInitializer.java
@@ -7,8 +7,7 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.ApplicationListener;
 import org.springframework.stereotype.Component;
 
-import com.cytech.trashtreasure.controller.HomeController;
-// import com.cytech.trashtreasure.controller.MainWindow;
+import com.cytech.trashtreasure.controller.LoginController;
 
 import net.rgielen.fxweaver.core.FxWeaver;
 
@@ -26,7 +25,7 @@ public class PrimaryStageInitializer implements ApplicationListener<StageReadyEv
     @Override
     public void onApplicationEvent(StageReadyEvent event) {
         Stage stage = event.stage;
-        Scene scene = new Scene(fxWeaver.loadView(HomeController.class), 640, 400);
+        Scene scene = new Scene(fxWeaver.loadView(LoginController.class), 640, 400);
         stage.setScene(scene);
         stage.setTitle(applicationTitle);
         stage.show();
