+++ src/main/resources/com/cytech/trashtreasure/controller/HomeView.fxml
@@ -1,59 +1,52 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
-<?import javafx.geometry.Insets?>
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.ButtonBar?>
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.control.PasswordField?>
-<?import javafx.scene.control.TextField?>
-<?import javafx.scene.layout.AnchorPane?>
-<?import javafx.scene.layout.HBox?>
-<?import javafx.scene.layout.VBox?>
+<!--
+  Copyright (c) 2015, 2019, Gluon and/or its affiliates.
+  All rights reserved. Use is subject to license terms.
 
-<VBox prefHeight="400.0" prefWidth="640.0" xmlns="http://javafx.com/javafx/19" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.cytech.trashtreasure.controller.HomeController">
+  This file is available and licensed under the following license:
+
+  Redistribution and use in source and binary forms, with or without
+  modification, are permitted provided that the following conditions
+  are met:
+
+  - Redistributions of source code must retain the above copyright
+    notice, this list of conditions and the following disclaimer.
+  - Redistributions in binary form must reproduce the above copyright
+    notice, this list of conditions and the following disclaimer in
+    the documentation and/or other materials provided with the distribution.
+  - Neither the name of Oracle Corporation nor the names of its
+    contributors may be used to endorse or promote products derived
+    from this software without specific prior written permission.
+
+  THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+  "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+  LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
+  A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+  OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+  SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+  LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+  DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+  THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+  (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+  OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+-->
+
+<?import javafx.scene.control.Label?>
+<?import javafx.scene.layout.AnchorPane?>
+<?import javafx.scene.layout.VBox?>
+<?import javafx.scene.text.Font?>
+
+<VBox fx:id="root" prefHeight="400.0" prefWidth="640.0" xmlns="http://javafx.com/javafx/19" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.cytech.trashtreasure.controller.HomeController">
   <children>
     <AnchorPane maxHeight="-1.0" maxWidth="-1.0" prefHeight="-1.0" prefWidth="-1.0" VBox.vgrow="ALWAYS">
-         <children>
-            <VBox layoutX="138.0" layoutY="116.0" prefHeight="118.0" prefWidth="364.0" spacing="10.0">
-               <children>
-                  <HBox prefHeight="21.0" prefWidth="421.0">
-                     <children>
-                        <Label text="Nom d'utilisateur">
-                           <HBox.margin>
-                              <Insets />
-                           </HBox.margin>
-                           <padding>
-                              <Insets right="10.0" />
-                           </padding>
-                        </Label>
-                        <TextField fx:id="usernameField" />
-                     </children>
-                  </HBox>
-                  <HBox prefHeight="27.0" prefWidth="267.0">
-                     <children>
-                        <Label text="Mot de passe " />
-                        <PasswordField fx:id="passwordField">
-                           <HBox.margin>
-                              <Insets left="31.0" />
-                           </HBox.margin>
-                        </PasswordField>
-                     </children>
-                  </HBox>
-               </children>
-               <padding>
-                  <Insets left="30.0" top="35.0" />
-               </padding>
-            </VBox>
-            <ButtonBar layoutX="388.0" layoutY="343.0" prefHeight="43.0" prefWidth="238.0">
-              <buttons>
-                  <Button fx:id="registerButton" mnemonicParsing="false" text="S'inscrire" />
-                <Button fx:id="loginButton" mnemonicParsing="false" text="Se connecter" />
-              </buttons>
-            </ButtonBar>
-         </children>
-         <padding>
-            <Insets right="10.0" />
-         </padding>
+      <children>
+        <Label alignment="CENTER" layoutX="228.0" layoutY="179.0" style="&#10;" text="Vous êtes connecté!" textAlignment="CENTER" textFill="#9f9f9f" wrapText="false">
+          <font>
+            <Font size="18.0" />
+          </font>
+        </Label>
+      </children>
     </AnchorPane>
   </children>
 </VBox>
